Single-step SDK install in docker container.

+++ tool/setup-dart.sh
@@ -26,9 +26,27 @@ then
   CHANNEL="be"
 fi
 
-mkdir -p "$1"
-cd "$1"
+# Create a temporary directory to extract the archive to.
+WORK_DIR=`mktemp -d`
+if [[ ! "$WORK_DIR" || ! -d "$WORK_DIR" ]]; then
+  echo "Could not create temporary directory."
+  exit 1
+fi
+
+# Download and extract Dart SDK
+cd "$WORK_DIR"
 curl -sS "https://storage.googleapis.com/dart-archive/channels/$CHANNEL/raw/$2/sdk/dartsdk-linux-x64-release.zip" >dartsdk.zip
 unzip -q dartsdk.zip
 rm -f dartsdk.zip
-./dart-sdk/bin/dart --disable-analytics
+
+# Move from temp to destination.
+DEST_PARENT=`dirname "$1"`
+mkdir -p "$DEST_PARENT"
+mv "$WORK_DIR/dart-sdk" "$1"
+
+# Initialize and set first run settings
+cd "$1"
+./bin/dart --disable-analytics
+
+# Cleanup
+rm -rf "$WORK_DIR"
